Added subscriptions to the users
