+++ account/src/main/java/com/github/smartoven/AccountApplication.java
@@ -1,0 +1,13 @@
+package com.github.smartoven;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AccountApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AccountApplication.class, args);
+    }
+
+}
