+++ account/src/main/java/com/github/smartoven/account/entity/UserRepository.java
@@ -1,0 +1,11 @@
+package com.github.smartoven.account.entity;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findUserByTelegramId(Long telegramId);
+
+    boolean existsByTelegramId(Long telegramId);
+}
