+++ account/src/main/java/com/github/smartoven/account/mapping/UserDto.java
@@ -1,0 +1,14 @@
+package com.github.smartoven.account.mapping;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDto {
+    private Long telegramId;
+}
