+++ account/src/main/java/com/github/smartoven/account/mapping/UserMapper.java
@@ -1,0 +1,25 @@
+package com.github.smartoven.account.mapping;
+
+import java.util.Collections;
+
+import com.github.smartoven.account.entity.User;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserMapper {
+
+    public User dtoToEntity(UserDto userDto) {
+        return User.builder()
+                .telegramId(userDto.getTelegramId())
+                .subscriptions(Collections.emptyList())
+                .build();
+    }
+
+    public UserViewModel entityToViewModel(User user) {
+        return new UserViewModel(
+                user.getId(),
+                user.getTelegramId(),
+                user.getSubscriptions()
+        );
+    }
+}
