+++ account/src/test/java/com/github/smartoven/account/AccountApplicationTests.java
@@ -1,0 +1,13 @@
+package com.github.smartoven.account;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AccountApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
