--- user/src/main/java/com/github/smartoven/UserApplication.java
@@ -1,13 +1,0 @@
-package com.github.smartoven;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserApplication.class, args);
-    }
-
-}
