--- user/src/main/java/com/github/smartoven/user/entity/User.java
@@ -1,40 +1,0 @@
-package com.github.smartoven.user.entity;
-
-import java.util.Collections;
-import java.util.List;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-
-import com.github.smartoven.user.entity.subscription.Subscription;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "users")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "telegram_id", unique = true)
-    private Long telegramId;
-
-    @OneToMany(fetch = FetchType.EAGER)
-    private List<Subscription> subscriptions = Collections.emptyList();
-}
