--- user/src/main/java/com/github/smartoven/user/mapping/UserDto.java
@@ -1,14 +1,0 @@
-package com.github.smartoven.user.mapping;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDto {
-    private Long telegramId;
-}
