--- user/src/main/java/com/github/smartoven/user/mapping/UserViewModel.java
@@ -1,15 +1,0 @@
-package com.github.smartoven.user.mapping;
-
-import java.util.List;
-
-import com.github.smartoven.user.entity.subscription.Subscription;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class UserViewModel {
-    private final Long id;
-    private final Long telegramId;
-    private final List<Subscription> subscriptions;
-}
